fix: bleu score location & assertion
