+++ README.md
@@ -147,6 +147,7 @@ Deterministic eval metrics
 | `python`                        | provided Python function validates the output                     |
 | `webhook`                       | provided webhook returns `{pass: true}`                           |
 | `rouge-n`                       | Rouge-N score is above a given threshold                          |
+| `bleu`                          | BLEU score is above a given threshold (default 0.5)               |
 | `levenshtein`                   | Levenshtein distance is below a threshold                         |
 | `latency`                       | Latency is below a threshold (milliseconds)                       |
 | `perplexity`                    | Perplexity is below a threshold                                   |
